requirements.txt
openai==1.76.0
pandas==2.2.3
pydantic==2.11.3
pyodbc==5.2.0
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.40
streamlit==1.44.1
streamlit_option_menu==0.4.0
